+++ tests/symbols.clj
@@ -1,3 +1,13 @@
+;; Problem:
+;; Pep finds usages of these symbols in paths,
+;; but from a var it doesn't find a symbol.
+;;
+;; Solution:
+;; Searching for usages of symbol or var should
+;; also include symbols.
+;;
+;; See https://github.com/pedrorgirardi/Pep/issues/68
+
 'clojure.core/map
 
 'example/foo